Sugar charts added to home page, works but needs improvement

+++ src/main/java/com/filipzyla/diabeticapp/backend/utility/SugarDefaultSettings.java
@@ -4,7 +4,7 @@ import com.filipzyla.diabeticapp.backend.enums.SugarUnits;
 
 public interface SugarDefaultSettings {
     SugarUnits DEFAULT_UNITS = SugarUnits.MILLI_GRAM;
-    double DEFAULT_HYPOGLYCEMIA = 70;
-    double DEFAULT_HYPERGLYCEMIA = 110;
-    double DEFAULT_HYPERGLYCEMIA_AFTER_MEAL = 150;
+    Integer DEFAULT_HYPOGLYCEMIA = 70;
+    Integer DEFAULT_HYPERGLYCEMIA = 110;
+    Integer DEFAULT_HYPERGLYCEMIA_AFTER_MEAL = 150;
 }
